Autonomous-Government-Scheme-Eligibility---Action-Agent-master/requirements.txt
# Government Scheme Eligibility Agent Requirements
# Core web framework and API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
python-multipart>=0.0.6
loguru>=0.7.2
aiofiles>=23.2.0

# Google Gemini AI integration (new package)
google-genai>=0.3.0

# Configuration and environment
python-dotenv>=1.0.0

# Data handling (optional)
pandas>=2.0.0
numpy>=1.24.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2